auth/pyproject.toml
[project]
name = "auth"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "cryptography==44.0.2",
    "pyjwt==2.10.1",
    "alembic>=1.15.2",
    "pydantic-settings>=2.9.0",
    "aiosqlite>=0.21.0",
    "psycopg2-binary>=2.9.10",
    "asyncpg>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.40",
    "passlib>=1.7.4",
    "bcrypt>=4.3.0",
]
